Add a Makefile and a coniguration file (config.sh.sample).
FossilOrigin-Name: bc2f94cd18e61b789742e759b919ffb5407308c89dd781186b22af17bc3f95ee

+++ build.sh
@@ -1,11 +1,18 @@
 #!/bin/sh
 
+set -e
+if [ ! -f config.sh ]
+then
+    echo 'Copy config.sh.sample to config.sh and edit it according to your requirements.'
+    exit 1
+fi
+
 for i in *.svg
 do
     bn=$(basename "$i" .svg)
     rm -rf "$bn"
     mkdir -p "$bn"
     cp "$i" "$bn"
-    (cd "$bn" && script fromsvg.log ../ficon.sh "$i" )
-    (cd "$bn" && script fromico.log ../cv.sh "$bn.ico" )
+    (source config.sh && cd "$bn" && script fromsvg.log ../ficon.sh "$i" )
+    (source config.sh && cd "$bn" && script fromico.log ../cv.sh "$bn.ico" )
 done
